Moved Deeplink Dialog to teams directory
