+++ src/dialogs/RootDialog.ts
@@ -8,7 +8,6 @@ import { ResetUserStateDialog } from "./examples/auth/ResetUserStateDialog";
 import { VSTSAPICallDialog } from "./examples/auth/VSTSAPICallDialog";
 import { VSTSAuthValidateUserDialog } from "./examples/auth/VSTSAuthValidateUserDialog";
 import { VSTSLogInDialog } from "./examples/auth/VSTSLogInDialog";
-import { DeeplinkDialog } from "./examples/basic/DeeplinkDialog";
 import { GetLastDialogUsedDialog } from "./examples/basic/GetLastDialogUsedDialog";
 import { HelloDialog } from "./examples/basic/HelloDialog";
 import { HelpDialog } from "./examples/basic/HelpDialog";
@@ -26,6 +25,7 @@ import { QuizQ1Dialog } from "./examples/moderate/QuizQ1Dialog";
 import { QuizQ2Dialog } from "./examples/moderate/QuizQ2Dialog";
 import { QuizQ3Dialog } from "./examples/moderate/QuizQ3Dialog";
 import { AtMentionDialog } from "./examples/teams/AtMentionDialog";
+import { DeeplinkDialog } from "./examples/teams/DeeplinkDialog";
 import { FetchRosterDialog } from "./examples/teams/FetchRosterDialog";
 import { ProactiveMsgTo1to1Dialog } from "./examples/teams/ProactiveMsgTo1to1Dialog";
 import { ProactiveMsgToChannelDialog } from "./examples/teams/ProactiveMsgToChannelDialog";
